+++ android/lib/src/main/cpp/CMakeLists.txt.in
@@ -26,6 +26,7 @@ add_library(androidapp SHARED
             ../../../../../wsd/ClientSession.cpp
             ../../../../../wsd/DocumentBroker.cpp
             ../../../../../wsd/COOLWSD.cpp
+            ../../../../../wsd/ClientRequestDispatcher.cpp
             ../../../../../wsd/RequestDetails.cpp
             ../../../../../wsd/Storage.cpp
             ../../../../../wsd/TileCache.cpp
