+++ wasm/Makefile.am
@@ -44,6 +44,7 @@ online_SOURCES = \
 	../wsd/ClientSession.cpp \
 	../wsd/DocumentBroker.cpp \
 	../wsd/COOLWSD.cpp \
+	../wsd/ClientRequestDispatcher.cpp \
 	../wsd/RequestDetails.cpp \
 	../wsd/Storage.cpp \
 	../wsd/TileCache.cpp \
